show coupons from the database n the index page

+++ app/controllers/CouponController.java
@@ -33,20 +33,12 @@ public class CouponController extends Controller {
 	 */
 	public static Result addCoupon() {
 
-		// String name = couponForm.bindFromRequest().get().name;
-		// double price = couponForm.bindFromRequest().get().price;
-		// Date created = couponForm.bindFromRequest().get().created;
-		// Date ending = couponForm.bindFromRequest().get().ending;
-		// String picture = couponForm.bindFromRequest().get().picture;
-		// String category = couponForm.bindFromRequest().get().category;
-		// String description = couponForm.bindFromRequest().get().description;
-		// Coupon.createCoupon(name, price, created, ending, picture, category,
-		// description);
-		// return ok(userIndex.render(null, null, name));
-
 		if (couponForm.hasErrors()) {
 			return redirect("/");// todo
 		}
+		
+		//TODO handle invalid inputs
+		
 		Coupon newc = couponForm.bindFromRequest().get();
 		long couponID = Coupon.createCoupon(newc.name, newc.price,
 				newc.dateCreated, newc.dateExpire, newc.picture, newc.category,
@@ -55,6 +47,9 @@ public class CouponController extends Controller {
 		return ok(userIndex.render(null, null, "Coupon \"" + newc.name
 				+ "\" added"));
 	}
+	
+	
+	
 
 	/**
 	 * Finds certain coupon
@@ -65,7 +60,7 @@ public class CouponController extends Controller {
 		return TODO;
 	}
 
-	public static Result viewCoupon(long id) {
+	public static Result viewCoupon(int id) {
 		Coupon coupon = Coupon.find(id);
 		return redirect("/");// todo//ok(viewCoupon.render(coupon, couponForm));
 	}
